Implemented API to retrieve 5-days forecast
Added ForecastProvider with entities and DTOs to retrieve 5-days/3-hours forecast for a city.
Added corresponding tests.

+++ XWeather/XWeather/Entities/Main.cs
@@ -3,8 +3,8 @@
     public class Main
     {
         public double temp { get; set; }
-        public int pressure { get; set; }
-        public int humidity { get; set; }
+        public double pressure { get; set; }
+        public double humidity { get; set; }
         public double temp_min { get; set; }
         public double temp_max { get; set; }
     }
